Add project tree and enable document deletion
The project tree was added to the session store containing diagram nodes for each project. Also, documents can now be deleted from a project. This change also includes updates to the user interface and related components needed to support these features. The project tree is displayed on the modeling page and can be manipulated by the user.