Add Layer functionality and improve permissions check
Added the functionality to create a new Layer document for a project. The Layer document will be used to group connection documents together based on their type. Updated the permissions check of the project tree nodes to include details like creating, editing, and deleting, instead of just checking for a DocumentType. Changed the padding for several icons to ensure consistency across all pages.